threaded data sending added
In that commit I've fixed the issue from from the previous one.

It was about adding background processes in order to receive th data
getting from sockets.

It was several troubles with integration that mecanic into the previous
remotePlayerFigth logic. It's about passing objectes of type sf::Socket
between threads. The problem is that it's not allowed to copy sf::Socket
object as well as we can not refer single socket object to the several
threads. So that, main tcpSocket class initialised globaly in main.cpp
and it is used only by the functions of the background threads.

So to receive data from socket you should get it from gRecvQueue as well
as to send a data. Thread-safety method of doing that s to call the void receiveData() or sendData(json data) (existed voids overloads)

I've also added a wait scene :)